Refactor and prevent from memory leaks
The early version of this gem wasn't optimized properly and was leaking
memory on each run, eventually leading to a Segmentation Fault and a
crash of the RubyVM...

This commit makes sure objects are released when they have to, and also
optimizes the memory allocations in general by doing them only once: in
the initializer/constructor.

Reproducing the crash on older versions is simple:

    # test.rb
    require 'blake2'

    1_000_000.times do
      Blake2.new(32, Blake2::Key.none).digest('abc', :to_hex)
    end

Running this script would eventually crash the VM after a few thousands
runs.

Props to Glenn Rempe (@grempe) for reporting this.

[Closes #3]